requirements.txt
alipay-sdk-python==3.2.0
aliyun-python-sdk-core==2.11.4
aliyun-python-sdk-core-v3==2.11.4
aliyun-python-sdk-kms==2.5.0
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
astroid==2.1.0
beautifulsoup4==4.6.3
billiard==3.3.0.23
bleach==3.0.2
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coverage==4.5.2
crcmod==1.7
cryptography==2.3.1
Django==2.0
django-celery==3.2.2
django-celery-beat==1.2.0
django-filter==2.0.0
django-filters==0.2.1
django-redis==4.9.0
django-simple-sso==0.14.0
django-timezone-field==3.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
filemagic==1.6
gunicorn==19.9.0
idna==2.7
image==1.5.27
isort==4.3.4
itsdangerous==1.1.0
kombu==3.0.37
lazy-object-proxy==1.3.1
mccabe==0.6.1
oss2==2.6.0
Pillow==5.2.0
pyasn1==0.4.4
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.7.2
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.5
pylint-plugin-utils==0.4
PyMySQL==0.9.2
python-crontab==2.3.5
python-dateutil==2.7.3
python-magic==0.4.15
pytz==2018.5
qrcode==6.0
raven==6.9.0
redis==2.10.6
requests==2.20.0
rsa==4.0
six==1.11.0
thrift==0.11.0
typed-ast==1.1.1
urllib3==1.24
vine==1.1.4
webencodings==0.5.1
webservices==0.7
wrapt==1.10.11
xlrd==1.2.0
xlwt==1.3.0
